pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.48.0",
    "gspread>=6.2.1",
    "pandas>=2.3.3",
    "streamlit>=1.54.0",
]
